+++ src/wai/annotations/isp/filter_metadata/__init__.py
@@ -1,0 +1,3 @@
+"""
+Package containing the filter-metadata ISP.
+"""