+++ src/wai/annotations/isp/filter_metadata/component/__init__.py
@@ -1,0 +1,1 @@
+from ._FilterMetadata import FilterMetadata