+++ CHANGES.rst
@@ -6,6 +6,7 @@ Changelog
 
 - Added `discard-invalid-images` ISP for removing corrupt images or annotations with no image attached
 - Added `batch-split` sub-command for splitting individual batches of annotations into subsets like train/test/val
+- Added `filter-metadata` ISP for filtering object detection
 
 
 0.1.6 (2022-05-11)